% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_tree_predict.R
\name{predict.prtree}
\alias{predict.prtree}
\title{Predict method for PRTree}
\usage{
\method{predict}{prtree}(object, newdata, ...)
}
\arguments{
\item{object}{Object of class inheriting from \code{"prtree"}}

\item{newdata}{A matrix with new values for the covariates.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list with the following arguments

 \item{yhat}{The predicted values.}

 \item{newdata}{The matrix with the covariates new values.}
}
\description{
Predicted values based on a prtree object.
}
